fixing small issue
