+++ .env
@@ -1,1 +1,1 @@
-API_USER_TOKEN=sample
+API_USER_TOKEN=eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIyNmE4NzMyMDM4M2ZlMzlhMTA3NDM4ZWRkMWI5YzBlNSIsInN1YiI6IjY1ZmQ1ZGE0NzcwNzAwMDE2MzA4NzkyNSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ._LeiSY9F_PLqdVc4kNAiLVIfos2gVt5eVFb-Wc1Nrj0
