+++ src/app/layout.js
@@ -1,7 +1,7 @@
 import { Inter } from "next/font/google"
 import "./globals.css"
 import NavBar from "@/components/layout/Navbar.client"
-import Footer from "@/components/footer"
+import Footer from "@/components/Footer/footer"
 
 const inter = Inter({ subsets: ["latin"] })
