+++ src/components/Movies/Movies.js
@@ -1,4 +1,3 @@
-import React from "react"
 import Card from "@/components/Card/Card"
 function Movies({ MoviesData }) {
   return (
